Move article markdown rendering from entity to view part
