--- apps/main/lib/main/entities/article.rb
@@ -1,20 +1,0 @@
-require "redcarpet"
-require "rom/struct"
-
-module Main
-  module Entities
-    class Article < ROM::Struct
-      def body_html
-        @body_html ||= markdown.render(body)
-      end
-
-      private
-
-      def markdown
-        @markdown ||= begin
-          Redcarpet::Markdown.new(Redcarpet::Render::HTML)
-        end
-      end
-    end
-  end
-end
